refactor: migrate to jest
