+++ .eslintrc
@@ -7,6 +7,7 @@
     "env": {
         "browser": true,
         "es2020": true,
+        "jest": true,
         "mocha": true,
         "node": true
     },
