+++ .gitignore
@@ -5,7 +5,6 @@ node_modules/
 dist/
 
 # test artifacts
-.nyc_output/
 artifacts/
 coverage/
 tmp/
