+++ package.json
@@ -76,13 +76,13 @@
     "grunt-cli": "^1.4.3",
     "grunt-contrib-clean": "^2.0.0",
     "grunt-contrib-nodeunit": "^4.0.0",
+    "jest": "^28.1.3",
     "jszip": "^3.10.0",
     "live-server": "^1.2.2",
     "load-grunt-tasks": "^5.1.0",
     "lunr": "^2.3.9",
     "marked": "^4.0.18",
     "mocha": "^10.0.0",
-    "nyc": "^15.1.0",
     "postcss": "^8.4.14",
     "postcss-cli": "^10.0.0",
     "postcss-import": "^14.1.0",
