+++ packages/atomizer/tests/fixtures/test.html
@@ -1,0 +1,1 @@
+<div class="D(b)"></div>
